V 1.00

+++ port_knocking/install.sh
@@ -4,18 +4,28 @@ source /etc/optimus/functions.sh
 output $OUTPUT_MODE
 output $OUTPUT_MODE "PROTECTION DU SERVEUR SSH AVEC UNE SEQUENCE DE PORT KNOCKING" "blue" 200 "port_knocking" 0
 
+if [ ! -f /etc/ufw/applications.d/ufw-webserver ]
+then
+	output $OUTPUT_MODE "Installation impossible : le Pare feu doit être installé préalablement" "red" 400 "port_knocking" 100
+	exit
+fi
+
+if [ -z $PORTKNOCKING_SEQUENCE ]
+then 
+	$PORTKNOCKING_SEQUENCE="port_knocking/knockd.conf"
+fi
+
 output $OUTPUT_MODE "Installation des paquets requis" "magenta" 200 "port_knocking" 20
 verbose apt-get -qq install knockd
 
 output $OUTPUT_MODE "Modification des fichiers de configuration" "magenta" 200 "port_knocking" 40
-verbose cp /etc/optimus/port_knocking/knockd.conf /etc/knockd.conf
+envsubst '${PORTKNOCKING_SEQUENCE}' < /etc/optimus/port_knocking/knockd.conf > /etc/knockd.conf
 if grep -q "Port 7822" /etc/ssh/sshd_config
 then
 	verbose sed -i 's/22/7822/g' /etc/knockd.conf
 fi
 verbose sed -i 's/START_KNOCKD=0/START_KNOCKD=1/g' /etc/default/knockd
-#Modification nécessaire pour rendre knockd compatible avec UFW mais qui devrait être intégrée nativement dans la prochaine version de knockd
-#verbose sed -i 's/ProtectSystem=full/ProtectSystem=true/g' /lib/systemd/system/knockd.service
+verbose sed -i 's/#KNOCKD_OPTS="-i eth0"/KNOCKD_OPTS="-i $(ip route get 8.8.8.8 | awk -- '{printf $5}')"/g' /etc/default/knockd
 
 output $OUTPUT_MODE "Redémarrage des services" "magenta" 200 "port_knocking" 60
 if ! grep -q "\[Install\]" /lib/systemd/system/knockd.service
@@ -26,12 +36,13 @@ verbose systemctl daemon-reload
 verbose systemctl restart knockd
 verbose systemctl --quiet enable knockd.service
 
-output $OUTPUT_MODE "Fermeture du port SSH" "magenta" 200 "port_knocking" 80
 if grep -q "Port 7822" /etc/ssh/sshd_config
 then
+	output $OUTPUT_MODE "Fermeture du port SSH 7822" "magenta" 200 "port_knocking" 80
 	verbose /sbin/ufw deny 7822
 else
+	output $OUTPUT_MODE "Fermeture du port SSH 22" "magenta" 200 "port_knocking" 80
 	verbose /sbin/ufw deny 22
 fi
 
-output $OUTPUT_MODE "Le serveur SSH a bien été sécurisé avec une sequence de port knocking !" "green" 200 "port_knocking" 100
+output $OUTPUT_MODE "La séquence de port knocking suivante a bien été configurée : $PORTKNOCKING_SEQUENCE" "green" 200 "port_knocking" 100
